refactor: rename directory themes -> groups
